rename: pages 폴더 추가 및 이동
