+++ src/pages/Home/Home.tsx
@@ -1,0 +1,11 @@
+import Button from '../../components/ui/button';
+
+function Home() {
+  return (
+    <>
+      <Button>버튼</Button>
+    </>
+  );
+}
+
+export default Home;
