+++ src/router.tsx
@@ -1,7 +1,7 @@
 import { createBrowserRouter } from 'react-router';
-import App from './App';
+import Home from './pages/Home/Home';
 import AppRoot from './AppRoot';
-import NotFound from './NotFound';
+import NotFound from './pages/NotFound/NotFound';
 
 /**
  * @link https://reactrouter.com/start/data/routing
@@ -11,7 +11,7 @@ const router = createBrowserRouter([
     path: '/',
     element: <AppRoot />,
     errorElement: <NotFound />,
-    children: [{ index: true, element: <App /> }],
+    children: [{ index: true, element: <Home /> }],
   },
 ]);
